+++ src/encoding/utf8.h
@@ -39,7 +39,7 @@ static constexpr uint32_t SHIFT = 6u;
 static constexpr uint32_t MASK = 0x3Fu; // 0011 1111
 
 // UTF-8 bytestring for given Unicode code point.
-uint32_t rune_to_bytes(uint32_t* s, rune r);
+uint8_t rune_to_bytes(uint8_t* s, rune r);
 
 // Read Unicode code point for the given (pre-validated) UTF-8 bytestring.
 uint32_t decode_unsafe(const uint8_t* str);
